Filter languages based on threshold (#21)
Hey Bu! 👋 

Thanks for building this project, it's super awesome and I'm sure as Hacktoberfest is under way lots of people are using it to find sweet new projects to contribute to!

This PR resolves #19 by adding a threshold of 10,000 lines of code for a language to be significant enough to be included in the list of language tags, but also includes the most popular language regardless. This should account for smaller repositories not having any languages above the 10k threshold, while still supporting larger projects.

Let me know if there's anything you'd like me to change 🙏 

✌️